vdr-mcli-plugin: just syncing - initial support for CAM-menus

+++ mcast/common/crc32.c
@@ -24,7 +24,7 @@ $Id: crc32.c,v 1.2 2006/01/02 18:24:04 rasc Exp $
 
 
 
-#include <sys/types.h>
+#include "defs.h"
 #include "crc32.h"
 
 
@@ -76,7 +76,7 @@ static u_long crc_table[256] = {
 	0x933eb0bb, 0x97ffad0c, 0xafb010b1, 0xab710d06, 0xa6322bdf, 0xa2f33668,
 	0xbcb4666d, 0xb8757bda, 0xb5365d03, 0xb1f740b4};
 
-u_long dvb_crc32 (char *data, int len)
+uint32_t dvb_crc32 (char *data, int len)
 {
 	register int i;
 	u_long crc = 0xffffffff;
